+++ src/java.desktop/windows/native/libawt/java2d/windows/GDIWindowSurfaceData.cpp
@@ -897,7 +897,7 @@ static void GDIWinSD_GetRasInfo(JNIEnv *env,
     }
     if (wsdo->lockFlags & SD_LOCK_LUT) {
         pRasInfo->lutBase =
-            (long *) wsdo->device->GetSystemPaletteEntries();
+            (jint *) wsdo->device->GetSystemPaletteEntries();
         pRasInfo->lutSize = 256;
     } else {
         pRasInfo->lutBase = NULL;
